SES-861: Update to v7.0.0
