+++ package.json
@@ -5,12 +5,14 @@
     "version": "kibana"
   },
   "authors": [
-  "David Moreno Lumbreras <dmorenolumb@gmail.com>"
+    "David Moreno Lumbreras <dmorenolumb@gmail.com>",
+    "Micahel Carius <michael@jsonar.com>",
+    "Joey Andres <joey@jsonar.com>"
   ],
-  "license": "Apache-2.0",
+  "license": "SEE LICENSE IN /opt/sonarfinder/sonarFinder/sonar_eula.txt",
   "repository": {
     "type": "git",
-    "url": "https://github.com/dlumbrer/kbn_searchtables"
+    "url": "https://github.com/jsonar/kbn_searchtables"
   },
   "scripts": {
     "lint": "eslint",
@@ -19,14 +21,5 @@
     "test:browser": "plugin-helpers test:browser",
     "build": "plugin-helpers build",
     "postinstall": "plugin-helpers postinstall"
-  },
-  "devDependencies": {
-    "@elastic/eslint-config-kibana": "0.0.2",
-    "@elastic/plugin-helpers": "^6.0.0",
-    "babel-eslint": "4.1.8",
-    "chai": "^3.5.0",
-    "eslint": "1.10.3",
-    "eslint-plugin-mocha": "1.1.0",
-    "lodash": "^4.17.5"
   }
 }
