+++ public/_table_vis.scss
@@ -1,0 +1,27 @@
+// SASSTODO: Update naming to BEM
+// This chart is actively being re-written to React and EUI
+// Putting off renaming to avoid conflicts
+.search-table-vis {
+  display: flex;
+  flex-direction: column;
+  flex: 1 0 100%;
+  overflow: auto;
+
+  .search-table-vis__searchBar {
+    flex-grow: 0;
+  }
+}
+
+.search-table-vis-container {
+  kbn-agg-table-group > .table > tbody > tr > td {
+    border-top: 0px;
+  }
+
+  .pagination-other-pages {
+    justify-content: flex-end;
+  }
+
+  .pagination-size {
+    display: none;
+  }
+}
