+++ public/index.scss
@@ -1,0 +1,10 @@
+@import 'src/legacy/ui/public/styles/styling_constants';
+
+// Prefix all styles with "tbv" to avoid conflicts.
+// Examples
+//   tbvChart
+//   tbvChart__legend
+//   tbvChart__legend--small
+//   tbvChart__legend-isLoading
+
+@import './table_vis';
