--- public/kbn_searchtables.html
@@ -1,42 +1,0 @@
-<div ng-controller="KbnSearchTablesVisController" class="kbn-searchtables">
-  <div class="kbn-searchtables-container">
-    <div class="navbar navbar-default subnav">
-      <div class="container-fluid">
-        <ul class="nav navbar-nav">
-          <li>
-            <div class="kuiLocalSearch" class="inputSearchContainer">
-              <div class="kuiLocalSearchAssistedInput" style="width:100%">
-                <input ng-model="config.searchKeyword"
-                  ng-keyup="$event.keyCode == 13"
-                  id="inputSearch_{{$id}}"
-                  placeholder="Search..."
-                  class="kuiLocalSearchInput ng-touched inputSearch"
-                  autocomplete="off"
-                  type="text">
-              </div>
-              <button aria-label="Search" class="kuiButton kuiButton--primary inputSearchButton">
-                <span aria-hidden="true" class="kuiIcon fa-search"></span>
-              </button>
-            </div>
-          </li>
-        </ul>
-      </div>
-    </div>
-    <kbn-agg-table-group
-      group="tableGroups"
-      export-title="vis.title"
-      per-page="vis.params.perPage"
-      sort="sort"
-      show-total="vis.params.showTotal"
-      total-func="vis.params.totalFunc"
-      ng-if="renderAgain && tableGroups">
-    </kbn-agg-table-group>
-    <div ng-if="!hasSomeRows && hasSomeRows !== null" class="table-vis-error">
-      <div class="euiSpacer euiSpacer--l"></div>
-      <h2 aria-hidden="true"><i aria-hidden="true" class="fa fa-meh-o"></i></h2>
-      <h4>No results found <br>
-        <a ng-click="resetSearch()">Reset Search</a>
-      </h4>
-    </div>
-  </div>
-</div>
