--- public/kbn_searchtables.less
@@ -1,63 +1,0 @@
-.kbn-searchtables {
-  display: flex;
-  flex-direction: column;
-  flex: 1 0 100%;
-}
-
-  .kbn-searchtables-error {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    flex: 1 0 100%;
-    text-align: center;
-  }
-
-  .kbn-searchtables-container {
-    kbn-agg-table-group > .table > tbody > tr > td {
-      border-top: 0px;
-    }
-
-    .pagination-other-pages {
-      justify-content: flex-end;
-    }
-
-    .pagination-size {
-      display: none;
-    }
-  }
-
-.inputSearchContainer
-{
-  width: 100%;
-}
-
-.kuiLocalSearch {
-  margin: 0.2em 1em;
-
-  .inputSearch {
-    width: 100%;
-    border-color: #ecf0f1;
-    border-width: 2px 0px 2px 2px;
-  }
-}
-
-.inputSearch:focus
-{
-  border-color: #b4bcc2;
-}
-
-.inputSearchButton, .cancelSearchButton
-{
-  height: 32px;
-  border-radius: 0px;
-}
-
-.kbn-searchtables-container {
-  ul.nav.navbar-nav {
-    width: 100%;
-
-    li {
-      width: 100%;
-    }
-  }
-}
