--- public/kbn_searchtables_params.html
@@ -1,44 +1,0 @@
-
-<div class="kuiSideBarSection">
-    <div class="form-group">
-      <label for="datatableVisualizationPerPage">Per Page</label>
-      <input type="number" ng-model="vis.params.perPage" class="form-control" id="datatableVisualizationPerPage">
-    </div>
-
-    <div class="checkbox">
-      <label>
-        <input type="checkbox" ng-model="vis.params.showMeticsAtAllLevels">
-        Show metrics for every bucket/level
-      </label>
-    </div>
-
-    <div class="checkbox">
-      <label>
-        <input type="checkbox" ng-model="vis.params.showPartialRows">
-        Show partial rows
-      </label>
-    </div>
-
-    <div class="checkbox">
-      <label>
-        <input type="checkbox" ng-model="metricsAtAllLevels" disabled>
-        Calculate metrics for every bucket/level
-      </label>
-    </div>
-
-    <div class="checkbox">
-      <label>
-        <input type="checkbox" ng-model="vis.params.showTotal">
-        Show total
-      </label>
-    </div>
-
-    <div>
-      <label for="datatableVisualizationTotalFunction">Total function</label>
-      <select ng-disabled="!vis.params.showTotal"
-              class="form-control"
-              ng-model="vis.params.totalFunc"
-              ng-options="x for x in totalAggregations" id="datatableVisualizationTotalFunction">
-      </select>
-    </div>
-  </div>
