+++ public/table_vis.html
@@ -1,0 +1,50 @@
+<div ng-controller="KbnSearchTableVisController" class="search-table-vis">
+  <div class="euiFlexGroup euiFlexGroup--gutterSmall euiFlexGroup--directionRow search-table-vis__searchBar">
+    <div class="euiFlexItem">
+      <div class="euiFormControlLayout euiFormControlLayout--fullWidth euiFormControlLayout--group">
+        <div class="euiFormControlLayout__childrenWrapper">
+          <input ng-model="config.searchKeyword"
+                 ng-keyup="$event.keyCode == 13"
+                 id="inputSearch_{{$id}}"
+                 placeholder="Search..."
+                 class="euiFieldText euiFieldText--fullWidth euiFieldText--inGroup"
+                 autocomplete="off"
+                 type="text">
+        </div>
+      </div>
+    </div>
+    <div class="euiFlexItem euiFlexItem--flexGrowZero">
+      <div class="euiFormControlLayout__childrenWrapper">
+        <button aria-label="Search" class="euiButton euiButton--primary">
+          <span aria-hidden="true" class="kuiIcon fa-search"></span>
+        </button>
+      </div>
+    </div>
+  </div>
+  <div ng-if="!hasSomeRows && hasSomeRows !== null" class="visError">
+    <div class="euiText euiText--extraSmall euiTextColor euiTextColor--subdued">
+      <icon type="'visualizeApp'" size="'m'" color="'subdued'"></icon>
+
+      <br>
+      <br>
+
+      <p
+        i18n-id="tableVis.vis.noResultsFoundTitle"
+        i18n-default-message="No results found">
+      </p>
+    </div>
+  </div>
+
+  <div ng-if="tableGroups" class="search-table-vis-container" data-test-subj="tableVis">
+    <kbn-agg-table-group
+      filter="vis.API.events.filter"
+      dimensions="dimensions"
+      group="tableGroups"
+      export-title="visState.title"
+      per-page="visState.params.perPage"
+      sort="sort"
+      show-total="visState.params.showTotal"
+      total-func="visState.params.totalFunc">
+    </kbn-agg-table-group>
+  </div>
+</div>
