+++ public/table_vis_params.html
@@ -1,0 +1,61 @@
+
+<div class="visEditorSidebar__section">
+  <div class="form-group">
+    <label
+      for="datatableVisualizationPerPage"
+      i18n-id="tableVis.params.perPageLabel"
+      i18n-default-message="Per Page"
+    ></label>
+    <input type="number" ng-model="editorState.params.perPage" class="form-control" id="datatableVisualizationPerPage">
+  </div>
+
+  <div class="checkbox">
+    <label>
+      <input type="checkbox" ng-model="editorState.params.showMetricsAtAllLevels" data-test-subj="showMetricsAtAllLevels">
+      <span
+        i18n-id="tableVis.params.showMetricsLabel"
+        i18n-default-message="Show metrics for every bucket/level"
+      ></span>
+    </label>
+  </div>
+
+  <div class="checkbox">
+    <label>
+      <input type="checkbox" ng-model="editorState.params.showPartialRows" data-test-subj="showPartialRows">
+      <span
+        i18n-id="tableVis.params.showPartialRowsLabel"
+        i18n-default-message="Show partial rows"
+      ></span>
+      &nbsp;
+      <icon-tip
+        content="::'tableVis.params.showPartialRowsTip' | i18n: {
+          defaultMessage: 'Show rows that have partial data. This will still calculate metrics for every bucket/level, even if they are not displayed.'
+        }"
+        position="'right'"
+      ></icon-tip>
+    </label>
+  </div>
+
+  <div class="checkbox">
+    <label>
+      <input type="checkbox" ng-model="editorState.params.showTotal">
+      <span
+        i18n-id="tableVis.params.showTotalLabel"
+        i18n-default-message="Show total"
+      ></span>
+    </label>
+  </div>
+
+  <div>
+    <label
+      for="datatableVisualizationTotalFunction"
+      i18n-id="tableVis.params.totalFunctionLabel"
+      i18n-default-message="Total function"
+    ></label>
+    <select ng-disabled="!editorState.params.showTotal"
+            class="form-control"
+            ng-model="editorState.params.totalFunc"
+            ng-options="x for x in totalAggregations" id="datatableVisualizationTotalFunction">
+    </select>
+  </div>
+</div>
